server/requirements.txt
Werkzeug ==0.11.3
Requests ==2.9.1
beautifulsoup4 ==4.4.1
python-dateutil ==2.4.2
html5lib ==0.999
